Выполнить работу над ошибками в ТЗ спринта №8
1. Изменить getPrioritizedTasks(), метод возвращает ArrayList
2. Исправить неправильную логику метода isTaskCross()
3. Исправить метод removeAllEpics()
4. Добавить тестовый класс TaskManagerTest
5. Изменить тестирование согласно ТЗ
