+++ src/manager/TaskManager.java
@@ -5,7 +5,6 @@ import data.Subtask;
 import data.Task;
 
 import java.util.ArrayList;
-import java.util.TreeSet;
 
 public interface TaskManager {
     void addTask(Task task);
@@ -48,5 +47,5 @@ public interface TaskManager {
 
     ArrayList<Task> getHistory();
 
-    TreeSet<Task> getPrioritizedTasks();
+    ArrayList<Task> getPrioritizedTasks();
 }
